Restyled by yapf

+++ src/pyate/term_extraction.py
@@ -28,9 +28,18 @@ Corpus = Union[str, Sequence[str]]
 class TermExtraction:
     # Utility function for defining patterns
     noun, adj, prep = (
-        {"POS": "NOUN", "IS_PUNCT": False},
-        {"POS": "ADJ", "IS_PUNCT": False},
-        {"POS": "ADP", "IS_PUNCT": False},
+        {
+            "POS": "NOUN",
+            "IS_PUNCT": False
+        },
+        {
+            "POS": "ADJ",
+            "IS_PUNCT": False
+        },
+        {
+            "POS": "ADP",
+            "IS_PUNCT": False
+        },
     )
 
     # Global settings for all instances of TermExtraction
@@ -49,13 +58,35 @@ class TermExtraction:
     # The pattern to check against for filtering term candidates (Ahrenberg, L. (2009). Term extraction : A Review Draft Version 091221. retrieved from https://www.semanticscholar.org/paper/Term-extraction-%3A-A-Review-Draft-Version-091221-Ahrenberg/f032f744813345b8b57f2780b4df951e4fa6d22a).
     patterns = [
         [adj],
-        [{"POS": {"IN": ["ADJ", "NOUN"]}, "OP": "*", "IS_PUNCT": False}, noun],
+        [{
+            "POS": {
+                "IN": ["ADJ", "NOUN"]
+            },
+            "OP": "*",
+            "IS_PUNCT": False
+        }, noun],
         [
-            {"POS": {"IN": ["ADJ", "NOUN"]}, "OP": "*", "IS_PUNCT": False},
+            {
+                "POS": {
+                    "IN": ["ADJ", "NOUN"]
+                },
+                "OP": "*",
+                "IS_PUNCT": False
+            },
             noun,
             prep,
-            {"POS": "DET", "OP": "?", "IS_PUNCT": False},
-            {"POS": {"IN": ["ADJ", "NOUN"]}, "OP": "*", "IS_PUNCT": False},
+            {
+                "POS": "DET",
+                "OP": "?",
+                "IS_PUNCT": False
+            },
+            {
+                "POS": {
+                    "IN": ["ADJ", "NOUN"]
+                },
+                "OP": "*",
+                "IS_PUNCT": False
+            },
             noun,
         ],
     ]
@@ -69,8 +100,8 @@ class TermExtraction:
             language = TermExtraction.language
         if language not in TermExtraction.nlps:
             TermExtraction.nlps[language] = spacy.load(
-                TermExtraction.config["spacy_model"], disable=["parser", "entity"]
-            )
+                TermExtraction.config["spacy_model"],
+                disable=["parser", "entity"])
         return TermExtraction.nlps[language]
 
     @staticmethod
@@ -83,12 +114,12 @@ class TermExtraction:
         if size is None:
             size = TermExtraction.config["DEFAULT_GENERAL_DOMAIN_SIZE"]
         if (language, size) not in TermExtraction.DEFAULT_GENERAL_DOMAINS:
-            TermExtraction.DEFAULT_GENERAL_DOMAINS[(language, size)] = pd.read_csv(
-                pkg_resources.resource_stream(
-                    __name__, f"default_general_domain.{language}.csv"
-                ),
-                nrows=size,
-            )
+            TermExtraction.DEFAULT_GENERAL_DOMAINS[(
+                language, size)] = pd.read_csv(
+                    pkg_resources.resource_stream(
+                        __name__, f"default_general_domain.{language}.csv"),
+                    nrows=size,
+                )
         return TermExtraction.DEFAULT_GENERAL_DOMAINS[(language, size)]
 
     @staticmethod
@@ -128,12 +159,12 @@ class TermExtraction:
         self.dtype = dtype
         if self.default_domain_size is None:
             self.default_domain_size = TermExtraction.config[
-                "DEFAULT_GENERAL_DOMAIN_SIZE"
-            ]
+                "DEFAULT_GENERAL_DOMAIN_SIZE"]
         if self.nlp is None:
             self.nlp = TermExtraction.get_nlp(self.language)
         if self.default_domain is None:
-            self.default_domain = TermExtraction.get_general_domain(self.language)
+            self.default_domain = TermExtraction.get_general_domain(
+                self.language)
         if self.max_word_length is None:
             self.max_word_length = TermExtraction.config["MAX_WORD_LENGTH"]
         if self.dtype is None:
@@ -164,8 +195,7 @@ class TermExtraction:
         """
         TermExtraction.config.update(new_settings)
         if not TermExtraction.config["model_name"].startswith(
-            TermExtraction.config["language"]
-        ):
+                TermExtraction.config["language"]):
             warnings.warn(
                 f"Model '{TermExtraction.config['model_name']}' and language '{TermExtraction.config['language']}' may not be compatible."
             )
@@ -203,22 +233,25 @@ class TermExtraction:
             def add_to_counter(matcher, doc, i, matches):
                 match_id, start, end = matches[i]
                 candidate = str(doc[start:end])
-                if TermExtraction.word_length(candidate) <= self.max_word_length:
+                if TermExtraction.word_length(
+                        candidate) <= self.max_word_length:
                     term_counter[candidate] += 1
 
             for i, pattern in enumerate(self.patterns):
-                new_matcher.add("term{}".format(i), [pattern], on_match=add_to_counter)
+                new_matcher.add("term{}".format(i), [pattern],
+                                on_match=add_to_counter)
 
             doc = self.nlp(document.lower(), disable=["parser", "ner"])
             new_matcher(doc)
         else:
             for end_index, (insert_order, original_value) in self.trie.iter(
-                document.lower()
-            ):
+                    document.lower()):
                 term_counter[original_value] += 1
         return term_counter
 
-    def count_terms_from_documents(self, seperate: bool = False, verbose: bool = False):
+    def count_terms_from_documents(self,
+                                   seperate: bool = False,
+                                   verbose: bool = False):
         """
         This is the main purpose of this class. Counts terms from the documents and returns a pandas Series.
         If self.corpus is a string, then it is identical to count_terms_from_document.
@@ -230,9 +263,9 @@ class TermExtraction:
             return self.__term_counts
 
         if type(self.corpus) is str:
-            self.__term_counts = pd.Series(
-                self.count_terms_from_document(self.corpus), dtype=self.dtype
-            )
+            self.__term_counts = pd.Series(self.count_terms_from_document(
+                self.corpus),
+                                           dtype=self.dtype)
             return self.__term_counts
         elif isinstance(self.corpus, collections.abc.Iterable):
             if seperate:
@@ -246,9 +279,8 @@ class TermExtraction:
                 if verbose:
                     pbar.update(1)
                 if seperate:
-                    term_counters.append(
-                        (tuple(counter_dict.keys()), tuple(counter_dict.values()))
-                    )
+                    term_counters.append((tuple(counter_dict.keys()),
+                                          tuple(counter_dict.values())))
                 else:
                     nonlocal term_counter
                     # update the cumulative/overall term_counter
@@ -279,14 +311,13 @@ class TermExtraction:
         if seperate:
 
             def counter_to_series(counter):
-                return pd.Series(data=counter[1], index=counter[0], dtype=self.dtype)
+                return pd.Series(data=counter[1],
+                                 index=counter[0],
+                                 dtype=self.dtype)
 
-            self.__term_counter = (
-                pd.DataFrame(data=map(counter_to_series, term_counters))
-                .fillna(0)
-                .astype(self.dtype)
-                .T
-            )
+            self.__term_counter = (pd.DataFrame(
+                data=map(counter_to_series, term_counters)).fillna(0).astype(
+                    self.dtype).T)
             return self.__term_counter
         else:
             self.__term_counter = pd.Series(
@@ -301,7 +332,6 @@ def add_term_extraction_method(extractor: Callable[..., pd.Series]):
     """
     Adds a method to the TermExtraction class. Rarely called by the user. For example, running add_term_extraction_method(combo_basic) would add the 'combo_basic' attribute to the TermExtraction class, so that TermExtraction(some_corpus).combo_basic could be called. This is for optimization, since if combo_basic is ran multiple times it would only count the number of terms once.
     """
-
     def term_extraction_decorated(self, *args, **kwargs):
         return extractor(
             self.corpus,
@@ -320,7 +350,5 @@ if __name__ == "__main__":
     wiki = pd.read_pickle(PATH_TO_GENERAL_DOMAIN)
     pmc = pd.read_pickle(PATH_TO_TECHNICAL_DOMAIN)
     print(
-        TermExtraction(pmc[:100]).count_terms_from_documents(
-            seperate=True, verbose=True
-        )
-    )
+        TermExtraction(pmc[:100]).count_terms_from_documents(seperate=True,
+                                                             verbose=True))
